Core Update

+++ webapp/src/forms.rs
@@ -4,6 +4,6 @@ define_form!(AuthResult, {
     id: String,
     display_name: String,
     auth_key: Option<String>,
-    roles: u64,
+    roles: u32,
     redirect_url: Option<String>
 });
